Refactor: inline attributes,  save several lines of code
