+++ memstore/src/lib.rs
@@ -1,5 +1,4 @@
-#[cfg(test)]
-mod test;
+#[cfg(test)] mod test;
 
 use std::collections::BTreeMap;
 use std::collections::HashMap;
