+++ openraft/src/engine/mod.rs
@@ -1,23 +1,15 @@
-#[allow(clippy::module_inception)]
-mod engine;
-
 mod command;
-#[cfg(test)]
-mod elect_test;
-#[cfg(test)]
-mod handle_vote_req_test;
-#[cfg(test)]
-mod handle_vote_resp_test;
-#[cfg(test)]
-mod initialize_test;
+mod engine_impl;
 mod log_id_list;
-#[cfg(test)]
-mod log_id_list_test;
-#[cfg(test)]
-mod purge_log_test;
-#[cfg(test)]
-mod testing;
+
+#[cfg(test)] mod elect_test;
+#[cfg(test)] mod handle_vote_req_test;
+#[cfg(test)] mod handle_vote_resp_test;
+#[cfg(test)] mod initialize_test;
+#[cfg(test)] mod log_id_list_test;
+#[cfg(test)] mod purge_log_test;
+#[cfg(test)] mod testing;
 
 pub(crate) use command::Command;
-pub(crate) use engine::Engine;
+pub(crate) use engine_impl::Engine;
 pub use log_id_list::LogIdList;
