+++ openraft/src/leader/mod.rs
@@ -1,4 +1,3 @@
-#[allow(clippy::module_inception)]
-mod leader;
+#[allow(clippy::module_inception)] mod leader;
 
 pub(crate) use leader::Leader;
