+++ openraft/src/metrics/mod.rs
@@ -6,14 +6,13 @@
 //!
 //! Metrics are observed on a running Raft node via the `Raft::metrics()` method, which will
 //! return a stream of metrics.
+
 mod raft_metrics;
 mod replication_metrics;
 mod wait;
 
-#[cfg(test)]
-mod replication_metrics_test;
-#[cfg(test)]
-mod wait_test;
+#[cfg(test)] mod replication_metrics_test;
+#[cfg(test)] mod wait_test;
 
 pub use raft_metrics::RaftMetrics;
 pub(crate) use replication_metrics::RemoveTarget;
