+++ openraft/src/vote/mod.rs
@@ -1,9 +1,7 @@
 mod leader_id;
-#[allow(clippy::module_inception)]
-mod vote;
+#[allow(clippy::module_inception)] mod vote;
 
 pub use leader_id::LeaderId;
 pub use vote::Vote;
 
-#[cfg(test)]
-mod leader_id_test;
+#[cfg(test)] mod leader_id_test;
