+++ rustfmt.toml
@@ -34,6 +34,7 @@ overflow_delimited_expr = true
 wrap_comments           = true
 comment_width           = 120
 max_width               = 120
+inline_attribute_width  = 80
 
 merge_derives           = false
